pyproject.toml
[build-system]
requires = [ "poetry>=0.12",]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "PyPPL"
version = "2.3.2"
description = "A Python PiPeLine framework"
authors = [ "pwwang <pwwang@pwwang.com>",]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/pwwang/PyPPL"
repository = "https://github.com/pwwang/PyPPL"
[[tool.poetry.packages]]
include = "pyppl/*.py"

[tool.poetry.dependencies]
python = "^3.6"
liquidpy = "*"
pyparam = "*"
filelock = "^3.0.0"
python-simpleconf = "*"
cmdy = "*"
pluggy = "^0.12"
colorama = "^0.4.1"
psutil = "^5.6"
transitions = "^0.6.0"
diot = "*"
attrs = "^19.3"
attr_property = "*"
python-varname = "*"
pyppl_report = "*"
pyppl_flowchart = "*"

[tool.poetry.dev-dependencies]
pytest = "*"
pytest-cov = "*"
jinja2 = "^2.0.0"
pyyaml = "^5.0.0"
faker = "^1.0.0"
graphviz = "^0.11"
toml = "^0.10"

[tool.poetry.scripts]
pyppl = "pyppl.console:main"
